requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
Deprecated==1.2.13
django==4.0.4
django-cleanup==6.0.0
django-environ==0.9.0
django-user-agents==0.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
hiredis==2.0.0
idna==3.3
mysqlclient==2.1.0
packaging==21.3
persiantools==3.0.1
Pillow==9.1.1
pyotp==2.6.0
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
requests==2.27.1
sqlparse==0.4.2
throttle==0.2.2
ua-parser==0.10.0
urllib3==1.26.9
user-agents==2.2.0
wrapt==1.14.1
